Stand-alone MNIST data set object
This is a variante implementation of our original loader. This code allows one to import this class file and use create a MNIST_Processing object from which all training, validation and testing vectors and labels as well as a modified aggregation of the data set specifically for our PyBrain feed-forward with back-propagation neural network can be obtained.